revert texture coordinates as ints instead of shorts (#46)
This reverts commit f994ffbe8c1b9480937b1b43f2ee225b8c362d95.

+++ include/PR/gbi.h
@@ -1118,7 +1118,7 @@
 typedef struct {
 	float		ob[3];	/* x, y, z */
 	unsigned short	flag;
-	int      		tc[2];	/* texture coord */
+	short		tc[2];	/* texture coord */
 	unsigned char	cn[4];	/* color & alpha */
 } Vtx_t;
 
@@ -1128,7 +1128,7 @@ typedef struct {
 typedef struct {
 	float		ob[3];	/* x, y, z */
 	unsigned short	flag;
-	int     		tc[2];	/* texture coord */
+	short		tc[2];	/* texture coord */
 	signed char	n[3];	/* normal */
 	unsigned char   a;      /* alpha  */
 } Vtx_tn;
